Address Copilot review on synchronous subscriptions
- Pass StartAt.subscriptionModelDefault() (not null) when the annotation
  processors register a @SynchronousSubscription, honoring the Subscribable
  contract even though the synchronous model ignores the start position.
- Use Math.toIntExact for the enrichment tail-read skip in both stream
  application services, so a stream version beyond Integer.MAX_VALUE fails loud
  instead of overflowing to a negative/wrong skip.
- Align the SynchronousSubscriptionModel null-check message with the
  writtenCloudEvents parameter name (blocking and reactor).
- Add the missing reactor SynchronousSubscriptionModel tests (duplicate id
  rejected, started handle returned) to match the blocking model.

+++ application/service/blocking/src/main/java/org/occurrent/application/service/blocking/generic/GenericApplicationService.java
@@ -146,7 +146,7 @@ public class GenericApplicationService<E> implements ApplicationService<E> {
                 // Re-read exactly the just-written tail so synchronous handlers get events enriched by the store
                 // (stream version and global position), then dispatch on this thread, inside the transaction.
                 int newEventCount = newEvents.size();
-                List<CloudEvent> writtenEnriched = eventStore.read(streamId, (int) writeResult.oldStreamVersion(), newEventCount).events().toList();
+                List<CloudEvent> writtenEnriched = eventStore.read(streamId, Math.toIntExact(writeResult.oldStreamVersion()), newEventCount).events().toList();
                 synchronousEventDispatcher.dispatch(writtenEnriched);
             }
