+++ application/service/reactor/src/main/java/org/occurrent/application/service/reactor/generic/GenericApplicationService.java
@@ -134,7 +134,7 @@ public class GenericApplicationService<E> implements ApplicationService<E> {
                                 }
                                 // Re-read exactly the just-written tail so synchronous handlers get events enriched by
                                 // the store (stream version and global position), then dispatch inside the transaction.
-                                return eventStore.read(streamId, (int) writeResult.oldStreamVersion(), newCloudEvents.size())
+                                return eventStore.read(streamId, Math.toIntExact(writeResult.oldStreamVersion()), newCloudEvents.size())
                                         .flatMap(enrichedStream -> enrichedStream.events().collectList())
                                         .flatMap(enriched -> synchronousEventDispatcher.dispatch(enriched).thenReturn(result));
                             });
